move BuildingGeometry into a separate file
Also rename it to BuildingGeometryForRadiation, because another already exists for thermal loads.
